second approach for fixing obspy.wav issue

+++ obspy/wav/core.py
@@ -140,7 +140,7 @@ def writeWAV(stream, filename, framerate=7000, rescale=False, width=4,
         if rescale:
             # optimal scale, account for +/- and the zero
             data = (2 ** (width * 8 - 1) - 1) * \
-                data.astype('f8') / abs(data).max()
+                data.astype('i8') / abs(data).max()
         data = np.require(data, dtype=WIDTH2DTYPE[width])
         w.writeframes(data.tostring())
         w.close()
